+++ src/report.c
@@ -108,10 +108,10 @@ static char *get_date_from_epoch (time_t *epoch_time,
   return buf;
 }
 
-static void print_file_hashes (uint64_t *hashes, int hash_count) {
+static void print_file_hashes (uint64_t *hashes, size_t hash_count) {
   printf("\t-> Hashes: [ ");
 
-  for (int i = 0; i < hash_count; i++) {
+  for (size_t i = 0; i < hash_count; i++) {
     printf("%016" PRIX64 " ", hashes[i]);
   }
 
@@ -155,7 +155,7 @@ void anu_print_report (anukrta_config *config,
              anu_time_microseconds_to_seconds(file->duration_us));
 
       if (config->verbose) {
-        print_file_hashes((hashes + (file_id * (size_t) config->segments)),
+        print_file_hashes((hashes + (file_id * config->segments)),
                           config->segments);
       }
     }
@@ -199,7 +199,7 @@ anu_report anu_generate_report (anu_file_q *files,
   /* Important to zero-initialize */
   anu_vector segment_results = {0};
   anu_vector_init(&segment_results, 32, sizeof(uint64_t));
-  const size_t segments = (size_t) config->segments;
+  const size_t segments = config->segments;
 
   for (size_t i = 0; i < file_count; i++) {
     for (size_t seg = 0; seg < segments; seg++) {
